Refactor reporting module
